Rearranged some of the files in the folder structure
Moved files that were in project directory to Source folder
