--- GAME3121_InfiniteEngine/InfiniteEngine/InfiniteEngine/AbilitiesInterface.h
@@ -1,16 +1,0 @@
-#pragma once
-
-
-class AbilitiesInterface
-{
-public:
-
-	AbilitiesInterface();
-
-
-
-	~AbilitiesInterface();
-};
-
-
-
