--- GAME3121_InfiniteEngine/InfiniteEngine/InfiniteEngine/Source.cpp
@@ -1,18 +1,0 @@
-#include "GameCodeApp.h"
-#include <iostream>
-
-using namespace std;
-
-int main()
-{
-	GameCodeApp* gGameCodeApp;
-	gGameCodeApp = new GameCodeApp();
-	
-	//gGameCodeApp->CheckMemory();
-
-
-
-
-	system("PAUSE");
-	return 0;
-}
