--- GAME3121_InfiniteEngine/InfiniteEngine/InfiniteEngine/SpecialAbilityComponent.cpp
@@ -1,11 +1,0 @@
-#include "SpecialAbilityComponent.h"
-
-
-
-SpecialAbilityComponent::SpecialAbilityComponent()
-{
-}
-
-SpecialAbilityComponent::~SpecialAbilityComponent()
-{
-}
