--- GAME3121_InfiniteEngine/InfiniteEngine/InfiniteEngine/SpecialAbilityComponent.h
@@ -1,8 +1,0 @@
-#pragma once
-class SpecialAbilityComponent
-{
-public:
-	SpecialAbilityComponent();
-	~SpecialAbilityComponent();
-};
-
