+++ GAME3121_InfiniteEngine/InfiniteEngine/Source/AbilitiesInterface.h
@@ -1,0 +1,16 @@
+#pragma once
+
+
+class AbilitiesInterface
+{
+public:
+
+	AbilitiesInterface();
+
+
+
+	~AbilitiesInterface();
+};
+
+
+
