+++ GAME3121_InfiniteEngine/InfiniteEngine/Source/Source.cpp
@@ -1,0 +1,18 @@
+#include "GameCodeApp.h"
+#include <iostream>
+
+using namespace std;
+
+int main()
+{
+	GameCodeApp* gGameCodeApp;
+	gGameCodeApp = new GameCodeApp();
+	
+	//gGameCodeApp->CheckMemory();
+
+
+
+
+	system("PAUSE");
+	return 0;
+}
