+++ GAME3121_InfiniteEngine/InfiniteEngine/Source/SpecialAbilityComponent.cpp
@@ -1,0 +1,11 @@
+#include "SpecialAbilityComponent.h"
+
+
+
+SpecialAbilityComponent::SpecialAbilityComponent()
+{
+}
+
+SpecialAbilityComponent::~SpecialAbilityComponent()
+{
+}
