+++ GAME3121_InfiniteEngine/InfiniteEngine/Source/SpecialAbilityComponent.h
@@ -1,0 +1,8 @@
+#pragma once
+class SpecialAbilityComponent
+{
+public:
+	SpecialAbilityComponent();
+	~SpecialAbilityComponent();
+};
+
