+++ GAME3121_InfiniteEngine/InfiniteEngine/Temp/x64/Release/InfiniteEngine.log
@@ -1,9 +1,10 @@
-﻿C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\Common7\IDE\VC\VCTargets\Microsoft.CppBuild.targets(376,5): warning MSB8004: Intermediate Directory does not end with a trailing slash.  This build instance will add the slash as it is required to allow proper evaluation of the Intermediate Directory.
-C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\Common7\IDE\VC\VCTargets\Microsoft.CppBuild.targets(377,5): warning MSB8004: Output Directory does not end with a trailing slash.  This build instance will add the slash as it is required to allow proper evaluation of the Output Directory.
+﻿C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\Common7\IDE\VC\VCTargets\Microsoft.CppBuild.targets(376,5): warning MSB8004: Intermediate Directory does not end with a trailing slash.  This build instance will add the slash as it is required to allow proper evaluation of the Intermediate Directory.
+C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\Common7\IDE\VC\VCTargets\Microsoft.CppBuild.targets(377,5): warning MSB8004: Output Directory does not end with a trailing slash.  This build instance will add the slash as it is required to allow proper evaluation of the Output Directory.
   Source.cpp
+  SpecialAbilityComponent.cpp
   Generating code
-  3 of 18 functions (16.7%) were compiled, the rest were copied from previous compilation.
+  1 of 18 functions ( 5.6%) were compiled, the rest were copied from previous compilation.
     0 functions were new in current compilation
     0 functions had inline decision re-evaluated but remain unchanged
   Finished generating code
-  InfiniteEngine.vcxproj -> C:\Users\Esk Eve\Documents\GitHub\Game3121-GameEngine\GAME3121_InfiniteEngine\InfiniteEngine\InfiniteEngine\..\Game\x64\Release\InfiniteEnginex64.exe
+  InfiniteEngine.vcxproj -> C:\Users\100978429\Documents\GitHub\Game3121-GameEngine\GAME3121_InfiniteEngine\InfiniteEngine\InfiniteEngine\..\Game\x64\Release\InfiniteEnginex64.exe
