+++ GAME3121_InfiniteEngine/InfiniteEngine/Temp/x64/Release/InfiniteEngine.tlog/InfiniteEngine.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.17763.0
-Release|x64|C:\Users\Esk Eve\Documents\GitHub\Game3121-GameEngine\GAME3121_InfiniteEngine\InfiniteEngine\|
+Release|x64|C:\Users\100978429\Documents\GitHub\Game3121-GameEngine\GAME3121_InfiniteEngine\InfiniteEngine\|
